Subset of growth data from the collaborative perinatal project (CPP)

Subset of growth data from the collaborative perinatal project (CPP).

cpp

Source

https://catalog.archives.gov/id/606622

Broman, Sarah. "The collaborative perinatal project: an overview." Handbook of longitudinal research 1 (1984): 185-227.

Examples

head(cpp)

#>   subjid agedays   wtkg htcm lencm      bmi siteid sexn  sex feedingn feeding
#> 1      1       1  4.621   55    55 15.27603      5    1 Male       90 Unknown
#> 2      1     123  8.760   NA    NA       NA      5    1 Male       90 Unknown
#> 3      1     366 14.500   79    79 23.23346      5    1 Male       90 Unknown
#> 4      2       1  3.345   51    51 12.86044      5    1 Male       90 Unknown
#> 5      2     123  4.340   NA    NA       NA      5    1 Male       90 Unknown
#> 6      2     366  8.400   73    73 15.76281      5    1 Male       90 Unknown
#>   gagebrth birthwt birthlen apgar1 apgar5 mage mracen mrace mmaritn  mmarit
#> 1      287    4621       55      8      9   21      5 White       1 Married
#> 2      287    4621       55      8      9   21      5 White       1 Married
#> 3      287    4621       55      8      9   21      5 White       1 Married
#> 4      280    3345       51      8      9   15      5 White       1 Married
#> 5      280    3345       51      8      9   15      5 White       1 Married
#> 6      280    3345       51      8      9   15      5 White       1 Married
#>   meducyrs sesn          ses parity gravida smoked mcignum preeclmp comprisk
#> 1       12   50       Middle      1       1      0       0        0     none
#> 2       12   50       Middle      1       1      0       0        0     none
#> 3       12   50       Middle      1       1      0       0        0     none
#> 4       NA   NA            .      0       0      1      35        0     none
#> 5       NA   NA            .      0       0      1      35        0     none
#> 6       NA   NA            .      0       0      1      35        0     none
#>   geniq sysbp diabp
#> 1    NA    NA    NA
#> 2    NA    NA    NA
#> 3    NA    NA    NA
#> 4    NA    NA    NA
#> 5    NA    NA    NA
#> 6    NA    NA    NA