Infer and attach attributes to a longitudinal growth study dataset

Infer attributes such as variable types of a longitudinal growth study and add as attributes to the dataset.

get_data_attributes(dat, meta = NULL, study_meta = NULL)

Arguments

dat
a longitudinal growth study dataset

meta
a data frame of meta data about the variables (a row for each variable)

study_meta
a single-row data frame or named list of meta data about the study (such as study description, etc.)

Details

attributes added: - subjectlevel_vars: vector of names of subject-level variables - longi_vars: vector of names of time-varying variables - time_vars: vector of names of measures of age - var_summ: data frame containing variable summaries with columns variable, label, type[subject id, time indicator, time-varying, constant], n_unique - subj_count: data frame of counts of records for each subject with columns subjid, n - n_subj: scalar containing total number of subjects - labels named list of variable labels - either populated by matching names with a pre-set list of labels (see hbgd_labels) or from a list provided from the meta argument - study_meta: data frame of meta data (if provided from the study_meta argument) - short_id: scalar containing the short unique identifier for the study (if study_meta is provided)

See also

Examples

cpp <- get_data_attributes(cpp)

#> `summarise_each()` is deprecated.
#> Use `summarise_all()`, `summarise_at()` or `summarise_if()` instead.
#> To map `funs` over all variables, use `summarise_all()`

str(attributes(cpp))

#> List of 4
#>  $ names    : chr [1:37] "subjid" "agedays" "wtkg" "htcm" ...
#>  $ row.names: int [1:1912] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ class    : chr "data.frame"
#>  $ hbgd     :List of 8
#>   ..$ labels           :List of 37
#>   .. ..$ subjid  : chr "Subject ID"
#>   .. ..$ agedays : chr "Age since birth at examination (days)"
#>   .. ..$ wtkg    : chr "Weight (kg)"
#>   .. ..$ htcm    : chr "Standing height (cm)"
#>   .. ..$ lencm   : chr "Recumbent length (cm)"
#>   .. ..$ bmi     : chr "BMI (kg/m**2)"
#>   .. ..$ waz     : chr "Weight for age z-score"
#>   .. ..$ haz     : chr "Length/height for age z-score"
#>   .. ..$ whz     : chr "Weight for length/height z-score"
#>   .. ..$ baz     : chr "BMI for age z-score"
#>   .. ..$ siteid  : chr "Investigational Site ID"
#>   .. ..$ sexn    : chr "Sex (numeric)"
#>   .. ..$ sex     : chr "Sex"
#>   .. ..$ feedingn: chr "Feeding practice (numeric)"
#>   .. ..$ feeding : chr "Feeding practice"
#>   .. ..$ gagebrth: chr "Gestational age at birth (days)"
#>   .. ..$ birthwt : chr "Birth weight (gm)"
#>   .. ..$ birthlen: chr "Birth length (cm)"
#>   .. ..$ apgar1  : chr "APGAR Score 1 min after birth"
#>   .. ..$ apgar5  : chr "APGAR Score 5 min after birth"
#>   .. ..$ mage    : chr "Maternal age at birth of child (yrs)"
#>   .. ..$ mracen  : chr "Maternal race (num)"
#>   .. ..$ mrace   : chr "Maternal race"
#>   .. ..$ mmaritn : chr "Mothers marital status (num)"
#>   .. ..$ mmarit  : chr "Mothers marital status"
#>   .. ..$ meducyrs: chr "Mother, years of education"
#>   .. ..$ sesn    : chr "Socioeconomic status of parent (num)"
#>   .. ..$ ses     : chr "Socioeconomic status of parent"
#>   .. ..$ parity  : chr "Maternal parity"
#>   .. ..$ gravida : chr "Maternal num pregnancies"
#>   .. ..$ smoked  : chr "Mom smoked during pregnancy?"
#>   .. ..$ mcignum : chr "Num cigarettes mom smoked per day"
#>   .. ..$ preeclmp: chr "Preeclampsia"
#>   .. ..$ comprisk: chr "Pregnancy complications/risk factors"
#>   .. ..$ geniq   : chr "Intelligence Quotient General Intelligence IQ"
#>   .. ..$ sysbp   : chr "Systolic Blood Pressure"
#>   .. ..$ diabp   : chr "Diastolic Blood Pressure"
#>   ..$ subjectlevel_vars: chr [1:24] "siteid" "sexn" "sex" "feedingn" ...
#>   ..$ longi_vars       : chr [1:11] "wtkg" "htcm" "lencm" "bmi" ...
#>   ..$ time_vars        : chr "agedays"
#>   ..$ var_summ         :Classes 'var_summ' and 'data.frame':	37 obs. of  5 variables:
#>   .. ..$ variable: chr [1:37] "subjid" "agedays" "wtkg" "htcm" ...
#>   .. ..$ label   : chr [1:37] "Subject ID" "Age since birth at examination (days)" "Weight (kg)" "Standing height (cm)" ...
#>   .. ..$ type    : chr [1:37] "subject id" "time indicator" "time-varying" "time-varying" ...
#>   .. ..$ vtype   : chr [1:37] "num" "cat" "num" "num" ...
#>   .. ..$ n_unique: int [1:37] 500 5 417 81 42 956 527 137 413 392 ...
#>   ..$ subj_count       :'data.frame':	500 obs. of  2 variables:
#>   .. ..$ subjid: int [1:500] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ n     : int [1:500] 3 4 1 1 1 1 4 5 5 2 ...
#>   ..$ n_subj           : int 500
#>   ..$ ad_tab           :Classes 'ad_tab' and 'data.frame':	5 obs. of  2 variables:
#>   .. ..$ agedays: int [1:5] 1 123 366 1462 2558
#>   .. ..$ n      : int [1:5] 500 460 428 137 387