Convert birth measurements to INTERGROWTH z-scores/centiles (generic)

Convert birth measurements to INTERGROWTH z-scores/centiles (generic)

igb_centile2value(gagebrth, p = 50, var = "lencm", sex = "Female")

igb_zscore2value(gagebrth, z = 0, var = "lencm", sex = "Female")

Arguments

gagebrth

gestational age at birth in days

p

centile(s) to convert (must be between 0 and 100)

var

the name of the measurement to convert ("lencm", "wtkg", "hcircm", "wlr")

sex

"Male" or "Female"

z

z-score(s) to convert

Note

For gestational ages between 24 and 33 weeks, the INTERGROWTH very early preterm standard is used.

References

International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project Villar, José et al. The Lancet, Volume 384, Issue 9946, 857-868

INTERGROWTH-21st very preterm size at birth reference charts. Lancet 2016 doi.org/10.1016/S0140-6736(16) 00384-6. Villar, José et al.

Examples

# get 99th centile for Male birth weights across some gestational ages
igb_centile2value(232:300, 99, var = "wtkg", sex = "Male")
#>  [1] 3.095594 3.134276 3.172468 3.210176 3.247402 3.284150 3.320425 3.356229
#>  [9] 3.391566 3.426440 3.460857 3.494817 3.528324 3.561383 3.593996 3.626168
#> [17] 3.657900 3.689198 3.720065 3.750502 3.780513 3.810103 3.839274 3.868027
#> [25] 3.896370 3.924301 3.951826 3.978946 4.005666 4.031988 4.057915 4.083450
#> [33] 4.108594 4.133353 4.157728 4.181721 4.205336 4.228576 4.251442 4.273939
#> [41] 4.296067 4.317829 4.339230 4.360270 4.380952 4.401281 4.421255 4.440879
#> [49] 4.460156 4.479087 4.497674 4.515922 4.533831 4.551403 4.568641 4.585547
#> [57] 4.602125 4.618375 4.634299 4.649901 4.665183 4.680145 4.694791 4.709122
#> [65] 4.723141 4.736849 4.750250 4.763343 4.776132