Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)

Convert postnatal growth measurements for preterm infants to INTERGROWTH z-scores/centiles (generic)

igprepost_centile2value(pmagedays, p = 50, var = "lencm", sex = "Female")

Arguments

pmagedays

postmenstrual age in days

p

centile(s) to convert (must be between 0 and 100)

var

the name of the measurement to convert ("lencm", "wtkg", "hcircm")

sex

"Male" or "Female"

References

Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project. Villar J, Giuliani F, Bhutta ZA, et al. Lancet Glob Health 2015; 3: e681–91.

Examples

# get median wtkg of Females from 27 to 64 weeks postmenstrual age
igprepost_centile2value(27:64 * 7, 50, var = "wtkg", sex = "Female")
#>  [1] 0.6132945 0.7571930 0.9149532 1.0851043 1.2660643 1.4562109 1.6539388
#>  [8] 1.8577023 2.0660448 2.2776189 2.4911968 2.7056754 2.9200755 3.1335381
#> [15] 3.3453183 3.5547771 3.7613733 3.9646542 4.1642472 4.3598506 4.5512257
#> [22] 4.7381896 4.9206075 5.0983868 5.2714710 5.4398351 5.6034803 5.7624303
#> [29] 5.9167277 6.0664307 6.2116106 6.3523488 6.4887355 6.6208672 6.7488458
#> [36] 6.8727764 6.9927671 7.1089273