List of WHO growth standard coefficients

A list of coefficients from the WHO for various pairs of growth standards that are used to compute quantiles and z-scores. The format is a list, where each element is a pairing of variables, e.g. "wtkg_agedays", "htcm_agedays", "bmi_agedays", "hcircm_agedays", "muaccm_agedays", "ss_agedays", "tsftmm_agedays", "wtkg_lencm""wtkg_htcm". Within each of these elements is a list for sex with names "Female" and "Male".

Source

0-5 years: http://www.who.int/childgrowth/software/en/

5-15 years: http://www.who.int/growthref/tools/en/

See also

Examples

head(who_coefs$htcm_age$Female$data)

#>      x l       m       s loh
#> 1858 0 1 49.1477 0.03790   L
#> 1859 1 1 49.3166 0.03783   L
#> 1860 2 1 49.4854 0.03776   L
#> 1861 3 1 49.6543 0.03770   L
#> 1862 4 1 49.8232 0.03763   L
#> 1863 5 1 49.9921 0.03756   L